+++ apps/iOS/EventsApp/Data/Persistence/CoreData/Models/H3Edge.swift
@@ -54,7 +54,7 @@ extension H3Edge: SyncTrackable {
     @NSManaged public var deletedAt: Date?
     @NSManaged public var syncStatusRaw: Int16
     @NSManaged public var lastCloudSyncedAt: Date?
-    @NSManaged public var schemaVersion: Int
+    @NSManaged public var schemaVersion: Int16
 
     // Computed:
     public var syncStatus: SyncStatus {
