+++ apps/iOS/EventsApp/Data/Persistence/CoreData/Models/PublicProfile.swift
@@ -40,7 +40,7 @@ extension PublicProfile: SyncTrackable {
     @NSManaged public var deletedAt: Date?
     @NSManaged public var syncStatusRaw: Int16
     @NSManaged public var lastCloudSyncedAt: Date?
-    @NSManaged public var schemaVersion: Int
+    @NSManaged public var schemaVersion: Int16
 
     // Computed:
     public var syncStatus: SyncStatus {
