+++ apps/iOS/EventsApp/Data/Persistence/CoreData/Models/UserSettings.swift
@@ -85,7 +85,7 @@ extension UserSettings: SyncTrackable {
     @NSManaged public var deletedAt: Date?
     @NSManaged public var syncStatusRaw: Int16
     @NSManaged public var lastCloudSyncedAt: Date?
-    @NSManaged public var schemaVersion: Int
+    @NSManaged public var schemaVersion: Int16
 
     // Computed:
     public var syncStatus: SyncStatus {
